refactor: styled components 파일명 변경
