+++ components/AddBar/AddBar.tsx
@@ -1,5 +1,6 @@
 import Image from "next/image";
-import * as Styled from "./StyledAddBar";
+import linkImg from "@/public/link.svg";
+import * as Styled from "./AddBar.styled";
 
 interface Props {
   isFixed: string;
@@ -10,12 +11,7 @@ const AddBar = ({ isFixed }: Props) => {
     <Styled.BackGround $isFixed={isFixed}>
       <Styled.Form>
         <label htmlFor="add">
-          <Image
-            src={"/link.svg"}
-            alt="링크 추가 이미지"
-            width={20}
-            height={20}
-          />
+          <Image src={linkImg} alt="링크 추가 이미지" width={20} height={20} />
         </label>
         <Styled.Input
           id="add"
