+++ components/FolderInfo/FolderInfo.style.ts
