+++ components/Footer/Footer.styled.ts
