+++ components/Modal/ModalInput.tsx
@@ -1,4 +1,4 @@
-import * as Styled from "./StyledModal";
+import * as Styled from "./Modal.styled";
 
 const ModalInput = () => {
   return <Styled.ModalInput placeholder="내용 입력" />;
