+++ components/Modal/ShareModal.tsx
@@ -1,7 +1,7 @@
 import { MouseEvent } from "react";
 import { ModalContentName } from "@/components";
 import ModalPortal from "@/lib/utils/Portal";
-import * as Styled from "./StyledModal";
+import * as Styled from "./Modal.styled";
 
 type closeModal = (e: MouseEvent) => void;
