+++ components/Navigator/NavBar.styled.ts
